remove straggler type files
