+++ libs/ui/src/components/atoms/inputLabel/inputLabel.tsx
@@ -1,7 +1,12 @@
 import Tooltip from '../tooltip/tooltip';
 import { ParMd } from '../typography';
-import { InputLabelType } from './inputLabelType';
 import { InputLabelContainer, RequiredAsterisk } from './inputLabelStyle';
+
+export type InputLabelType = {
+  required?: boolean;
+  info?: string;
+  children: React.ReactChild;
+};
 
 const InputLabel = ({ required, info, children = 'label' }: InputLabelType) => {
   return (
