--- libs/ui/src/components/atoms/inputLabel/inputLabelType.ts
@@ -1,5 +1,0 @@
-export type InputLabelType = {
-  required?: boolean;
-  info?: string;
-  children: React.ReactChild;
-};
