--- libs/ui/src/components/atoms/tooltip/tooltipTypes.ts
@@ -1,7 +1,0 @@
-export type TooltipProps = {
-  content?: string | React.ReactNode;
-  side?: 'top' | 'right' | 'bottom' | 'left';
-  triggerEl?: React.ReactNode;
-  offset?: number;
-  delay?: number;
-};
